refactor: flatten server agent directory
