+++ apps/server/src/agent/ai-sdk-agent.ts
@@ -1,20 +1,20 @@
 import { AGENT_LIMITS } from '@browseros/shared/constants/limits'
 import type { BrowserContext } from '@browseros/shared/schemas/browser-context'
 import { stepCountIs, ToolLoopAgent, type UIMessage } from 'ai'
-import type { Browser } from '../../browser/browser'
-import type { KlavisClient } from '../../lib/clients/klavis/klavis-client'
-import { logger } from '../../lib/logger'
-import { isSoulBootstrap, readSoul } from '../../lib/soul'
-import { buildFilesystemToolSet } from '../../tools/filesystem/build-toolset'
-import { buildMemoryToolSet } from '../../tools/memory/build-toolset'
-import type { ToolRegistry } from '../../tools/tool-registry'
-import { CHAT_MODE_ALLOWED_TOOLS } from '../chat-mode'
-import { buildSystemPrompt } from '../prompt'
-import type { ResolvedAgentConfig } from '../types'
+import type { Browser } from '../browser/browser'
+import type { KlavisClient } from '../lib/clients/klavis/klavis-client'
+import { logger } from '../lib/logger'
+import { isSoulBootstrap, readSoul } from '../lib/soul'
+import { buildFilesystemToolSet } from '../tools/filesystem/build-toolset'
+import { buildMemoryToolSet } from '../tools/memory/build-toolset'
+import type { ToolRegistry } from '../tools/tool-registry'
+import { CHAT_MODE_ALLOWED_TOOLS } from './chat-mode'
 import { createCompactionPrepareStep } from './compaction'
 import { buildMcpServerSpecs, createMcpClients } from './mcp-builder'
+import { buildSystemPrompt } from './prompt'
 import { createLanguageModel } from './provider-factory'
 import { buildBrowserToolSet } from './tool-adapter'
+import type { ResolvedAgentConfig } from './types'
 
 export interface AiSdkAgentConfig {
   resolvedConfig: ResolvedAgentConfig
