+++ apps/server/src/agent/format-message.ts
