+++ apps/server/src/api/services/chat-service.ts
@@ -7,9 +7,9 @@
 import { mkdir } from 'node:fs/promises'
 import path from 'node:path'
 import { createAgentUIStreamResponse, type UIMessage } from 'ai'
-import { AiSdkAgent } from '../../agent/tool-loop/ai-sdk-agent'
-import { formatUserMessage } from '../../agent/tool-loop/format-message'
-import type { SessionStore } from '../../agent/tool-loop/session-store'
+import { AiSdkAgent } from '../../agent/ai-sdk-agent'
+import { formatUserMessage } from '../../agent/format-message'
+import type { SessionStore } from '../../agent/session-store'
 import type { ResolvedAgentConfig } from '../../agent/types'
 import type { Browser } from '../../browser/browser'
 import type { KlavisClient } from '../../lib/clients/klavis/klavis-client'
